+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.55:
+* Add FlushingTestModule.bindService() method that takes classnames.
 * Rework PersistenceTestModule constructor to pass databasePrefix as an optional parameter.
 * Remove PersistenceTestModule.registerUserTransaction() as it is unused.
 * Pass the tables to clean as a parameter to the constructor.